% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxnodf.R
\name{maxnodf}
\alias{maxnodf}
\title{Compute NODF-maximising graph}
\usage{
maxnodf(NodesA, NodesB, Edges, quality = 0)
}
\arguments{
\item{NodesA}{The number of rows in the desired network.}

\item{NodesB}{The number of columns in the desired network.}

\item{Edges}{The number of Edges in the desired network.}

\item{quality}{An optional quality parameter to control the
tradeoff between computation time and result quality.}
}
\value{
Approximation of the NODF maximising graph.
}
\description{
Compute NODF-maximising graph
}
\examples{
maxnodf(14, 13, 52)
maxnodf(14, 13, 52, 2)
}
