% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxnodf.R
\name{maxnodf}
\alias{maxnodf}
\title{Compute NODF-maximising network}
\usage{
maxnodf(web, quality = 0)
}
\arguments{
\item{web}{Either a numeric matrix describing a mutualistic network
or a numeric vector of length 3 of the form
web = c(#Rows, #Columns, #Links).}

\item{quality}{An optional quality parameter to control the
tradeoff between computation time and result quality.}
}
\value{
Nestedness and graph-matrix of the NODF maximising network.
}
\description{
Compute NODF-maximising network
}
\examples{
maxnodf(matrix(1.0, 12, 10))
maxnodf(c(14, 13, 52), 2)
}
